Move Tune Menu to its own file
